chore: Move all cli scripts into lib folder.
