+++ .eslintignore
@@ -1,5 +1,5 @@
 node_modules/*
-dist
+lib-dist
 test/fixtures
 coverage
 esdocs
