+++ .gitignore
@@ -1,5 +1,5 @@
 # compiled es6 code
-dist/
+lib-dist/
 coverage
 node_modules
 npm-debug.log
