+++ bin/reptar-debug.js
@@ -1,6 +1,12 @@
 #!/usr/bin/env node
+/* eslint-disable global-require */
 
 require('babel-polyfill');
 require('babel-register');
 
-require('./index');
+const path = require('path');
+
+require('./index')({
+  log: require('../lib/log').default,
+  libPath: path.join(__dirname, '..', 'lib'),
+});
