+++ bin/reptar.js
@@ -1,5 +1,11 @@
 #!/usr/bin/env node
+/* eslint-disable global-require */
 
 require('babel-polyfill');
 
-require('./index');
+const path = require('path');
+
+require('./index')({
+  log: require('../lib-dist/log').default,
+  libPath: path.join(__dirname, '..', 'lib-dist'),
+});
