+++ lib/cli/build.js
@@ -1,5 +1,5 @@
-import log from '../lib/log';
-import Reptar from '../lib';
+import log from '../log';
+import Reptar from '../index';
 
 export default async function build(options) {
   const id = log.startActivity('building\t\t\t');
