+++ lib/cli/clean.js
@@ -1,5 +1,5 @@
-import log from '../lib/log';
-import Reptar from '../lib';
+import log from '../log';
+import Reptar from '../index';
 
 export default function () {
   const reptar = new Reptar();
