+++ lib/cli/init.js
@@ -3,8 +3,8 @@ import moment from 'moment';
 import path from 'path';
 import inquirer from 'inquirer';
 import spawn from 'cross-spawn';
-import log from '../lib/log';
-import Constants from '../lib/constants';
+import log from '../log';
+import Constants from '../constants';
 
 const dateNowFormatted = moment().format('YYYY-M-D');
