+++ lib/cli/new.js
@@ -3,9 +3,9 @@ import path from 'path';
 import moment from 'moment';
 import fs from 'fs-extra';
 import inquirer from 'inquirer';
-import Config from '../lib/config';
-import Url from '../lib/url';
-import log from '../lib/log';
+import Config from '../config';
+import Url from '../url';
+import log from '../log';
 
 const newTypes = {
   file: {
