+++ lib/cli/serve.js
@@ -1,8 +1,8 @@
 import Hapi from 'hapi';
 import inert from 'inert';
 import path from 'path';
-import Config from '../lib/config';
-import log from '../lib/log';
+import Config from '../config';
+import log from '../log';
 
 export default async function serve() {
   const config = new Config();
