+++ lib/cli/watch.js
@@ -5,9 +5,9 @@ import Hapi from 'hapi';
 import Boom from 'boom';
 import inert from 'inert';
 import chokidar from 'chokidar';
-import Constants from '../lib/constants';
-import log from '../lib/log';
-import Reptar from '../lib';
+import Constants from '../constants';
+import log from '../log';
+import Reptar from '../index';
 
 // Given an obj it'll prune any properites that start with `_`.
 function prunePrivateProperties(obj, isPrivate = (val, key) => key[0] === '_') {
